+++ golboogi/src/main/java/com/gf/golboogi/repository/ReviewDao.java
@@ -24,7 +24,7 @@ public interface ReviewDao {
 
 	List<ReviewDto> otherview(String fieldName);
 
-	Double ratingView(String fieldName);
+	Float ratingView(String fieldName);
 
 	List<RecentReviewVO> recentReview(int count);
